Couple of things here:
Removing semi-functioning pcap generation.

Spliting lorawan-frame-header into lorawan-frame-header-uplink and lorawan-frame-header-downlink, as there is a difference in format in the uplink and downlink versions.

Adding Class B bit (as specified in the LoRaWAN specification) that signified to the Network Server that a device is operating in Class B mode. The Network Server has been changed to use this bit.

Class B devices can now operate in beacon-less mode for 2 hours, it is assumed that the device can use it's own on-board clock to calculate an accurate timestamp for this amount of time.

Fixed a bug where the AES key was not being initialised properly.

Removed a failing example.

Updated tests to reflect changes made to the frame headers.